Make base class for SingleDatasetPerformance checks (tabular & nlp)
