+++ deepchecks/tabular/suite.py
@@ -17,7 +17,6 @@ import pandas as pd
 
 from deepchecks.core import DatasetKind
 from deepchecks.core.check_result import CheckFailure
-from deepchecks.core.errors import DeepchecksNotSupportedError
 from deepchecks.core.suite import BaseSuite, SuiteResult
 from deepchecks.tabular._shared_docs import docstrings
 from deepchecks.tabular.base_checks import ModelOnlyCheck, SingleDatasetCheck, TrainTestCheck
@@ -144,7 +143,3 @@ class Suite(BaseSuite):
                 results.append(CheckFailure(check, exp))
 
         return SuiteResult(self.name, results)
-
-    @classmethod
-    def _get_unsupported_failure(cls, check, msg):
-        return CheckFailure(check, DeepchecksNotSupportedError(msg))
